training_3d/requirements.txt
numpy
pillow
requests
torch
torchvision
scikit-learn
sqlalchemy
psycopg2-binary
pyyaml
pydantic
pandas
mlflow
tifffile
matplotlib
